# swagger/api.yaml
openapi: 3.0.0
info:
  description: "This is the static API documentation for the project called medical-metadata which deals as a proof of concept for the master thesis of Bence Slajchó @Alpen-Adria-University-Klagenfurt. Supervisors: O.Univ.-Prof. Dipl.-Ing. Dr. Johann Eder; Dr.phil. Volodymyr Shekhovtsov"
  version: "1.3.0"
  title: "Medical Metadata"
  contact:
    email: "slajchobence@gmail.com"
...
servers:
  - url: localhost:3001
    description: development server
paths:
...
  /api/datainfo/collectionnames:
    get:
      tags:
      - dataset
      summary: "returns the names of the collecitons which were generated"
      description: "the returned data contains an array of colelction names"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCollectionNamesGet"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # Data collections deletion
  /api/datainfo/collections:
    delete:
      tags:
      - "dataset"
      summary: "delete all collection raw data"
      description: "regardless if the collection was generated by the system or uploaded from an external source, all collections will be deleted"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCollectionsDelete"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # Data collections deletion for a specific collection
  /api/datainfo/collection:
    delete:
      tags:
      - "dataset"
      summary: "deletes collection data for a specific collection"
      description: "a collection's name is passed to this route and it will delete all the corresponding data for it"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                collectionName:
                  type: "string"
                  example: "raw_medical_data_1"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCollectionDelete"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # Data collections deletion for a specific collection
...
  /api/loinc-parent-groups:
    get:
      tags:
      - search
      summary: returns all the loinc parent groups
      description: "without any filtering"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLoincParentGroups"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # relevant LOINC parent groups
  /api/loinc-relevant-parent-groups:
    get:
      tags:
      - search
      summary: returns only the parent groups which are in the researched domain
      description: "filters on an additional flag if the parent group is used in the project or not"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLoincParentGroups"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # LOINC grups which belong to a parent group
  /api/loinc-groups:
    get:
      tags:
      - search
      summary: "returns loinc groups which belong to a parent group"
      description: ""
      parameters:
        - in: query
          name: parentGroupIdList
          required: true
          schema:
            type: array
            example: [LG74-7, LG97-8, LG78-8]
            items:
              type: string
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLoincGroups"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # LOINC terms
  /api/loinc-terms:
    get:
      tags:
      - search
      summary: "returns loinc terms which belong to the selected group"
      description: ""
      parameters:
        - in: query
          name: parentGroupIdList
          required: true
          schema:
            type: array
            example: [ "LG35818-0","LG34479-2","LG34487-5" ]
            items:
              type: string
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLoincTerms"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # LOINC concept informations
  /api/loinc-fsn:
    get:
      tags:
      - search
      summary: "returns information of the loinc concept based on the loinc code"
      description: "Information is retrieved from the official loinc fhir api, yet there is an interceptor in the backend. This interceptor checks if the scaling of ORD is truly an ORD scaling or is it handled as NOM on meta level. If so, the SCALE_TYP is returned as NOM (even if the corresponding LOINC code has initially ORD scaling)"
      parameters:
        - in: query
          name: loinc_code
          required: true
          schema:
            type: string
            example: 32167-9
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLoincFsn"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # LOINC values ranges for Qt
...
  /api/loinc-part-value:
    get:
      tags:
      - "search"
      summary: "Lists all the available LOINC part values from the system. Only those get listed which are exisitng in the metadata level in a form of a LOINC code as collection desctiptor."
      description: "This is a general route for any LOINC part values. LOINC's fully specified name consists of the LOINC code and six parts. These six parts are in sequential order: Component (1), Property(2), Time aspect(3), System(4), Scale(5), Method(6). 
        For this route the sequnetial number is required (1 / 2 / 3 / 4 / 5/ 6) and in return all distinct values are returned which are in the system as collection descriptors." 
      parameters:
        - in: query
          name: loincPartSelected
          required: true
          schema:
            type: string
            example: "2"
            enum: ["1", "2", "3", "4", "5", "6"]
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLoincPartValue"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/collections-to-loinc:
    get:
      tags:
      - "search"
      summary: "Lists collections which are described by a set of LOINC codes."
      parameters:
        - in: query
          name: loincCodes
          required: true
          schema:
            type: array
            items: 
              type: string
              example: ["5794-3"]
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCollectionToLoinc"
        "400":
          description: "no loinc code"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCollectionToLoincNoCode"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/collection-to-loinc-part-value:
    get:
      tags:
      - "search"
      summary: "Lists all collections which satisfy the query."
      description: "If the search value of any LOINC part refers to a LOINC Part Value in the metadata level, a list of collections is returned which contain LOINC codes where the LOINC Part Values equals to the search value." 
      parameters:
        - in: query
          name: component
          required: false
          schema:
            type: string
            example: "ph"
        - in: query
          name: property
          required: false
          schema:
            type: string
            example: "PrThr"
        - in: query
          name: time_aspect
          required: false
          schema:
            type: string
            example: "Pt"
        - in: query
          name: system
          required: false
          schema:
            type: string
            example: "Urine"
        - in: query
          name: scale
          required: false
          schema:
            type: string
            example: "Nom"
        - in: query
          name: method
          required: false
          schema:
            type: string
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCollectionToLoincPartValue"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
...
  /api/metalevel-loinc-codes:
    get:
      tags:
      - "search"
      summary: "List of all LOINC codes including the Long Common Names"
      description: "The list of the LOINC codes consist of those which are used on the metadata level as collection desctiptors.
        In this route all used LOINC codes are returned including their correcponding Long Common Name" 
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMetalevelLoincCodes"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/loinc-code-operators:
    get:
      tags:
      - "composite descriptor"
      summary: "List of all possible opeartors which can be choosed from at the creation of a composite collection descriptor"
      description: "The list of such operators contains two possible elements: 'ALL' and 'ANY'. These two opeartors are implemented on system level.
        ANY would apply at the search in the following way: if the composite key contains LOINC code 1 and LOINC code 2, then any collection would be returned by the query which contain either LOINC code 1 OR LOINC code 2 (maybe both). Example: 'Find all collections which contain any primary diagnosis'
        ALL would apply at the search in the following way: if the composite key contains LOINC code 1 and LOINC code 2, then only those collections will be returned which contain both LOINC code 1 AND LOINC code 2. Example: 'Find all collections which contain BMI related information'" 
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLoincCodeOperators"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
  /api/collections-to-composite-descriptors:
    get:
      tags:
      - "composite descriptor"
      summary: "Lists collections which satisfy the composite collection desctiptor"
      description: "The collections are returned which satisfy the composie collection descriptors (pre-requisite: an admin has defined such composite descriptors beforehand from which the user can then choose. Info about the dectiptor is provided).
        ANY would apply at the search in the following way: if the composite key contains LOINC code 1 and LOINC code 2, then any collection would be returned by the query which contain either LOINC code 1 OR LOINC code 2 (maybe both). Example: 'Find all collections which contain any primary diagnosis'
        ALL would apply at the search in the following way: if the composite key contains LOINC code 1 and LOINC code 2, then only those collections will be returned which contain both LOINC code 1 AND LOINC code 2. Example: 'Find all collections which contain BMI related information'" 
      parameters:
        - in: query
          name: operator
          required: false
          schema:
            type: string
            example: "ANY"
            enum: ["ANY", "ALL"]
        - in: query
          name: participatingCodes
          required: false
          schema:
            type: array
            items:
              type: string
              example: ["5794-3"]
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCollectionsToCompositeDescriptors"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
...
components:
  schemas:
...
    ApiResponseLoincParentGroups:
      type: "array"
      items:
        type: "object"
        properties:
          parent_group_id:
            type: string
            example: "LG74-7"
          parent_group:
            description: the name of the current parent group
            type: string
            example: "Urine<SAME:Comp|Prop|Sys><ROLLUP:Tm|Meth>"
    ApiResponseLoincGroups:
      type: "array"
      items:
        type: "object"
        properties:
          parent_group_id:
            type: string
            example: "LG74-7"
          group_id:
            type: string
            example: "LG36551-6"
          group_name:
            type: string
            example: "11-Dehydro thromboxane beta 2/Creatinine|MRto|Urine"
    ApiResponseLoincTerms:
      type: "array"
      items:
        type: "object"
        properties:
          loinc_code:
            type: string
            example: "32167-9"
          long_common_name:
            type: string
            example: "Clarity of Urine"
          group_id:
            type: string
            example: "LG35818-0"
...
    ApiResponseLoincFsn:
      type: "object"
      properties:
        COMPONENT:
          type: string
          example: "Clarity"
        PROPERTY:
          type: string
          example: "Type"
        SCALE_TYP:
          type: string
          example: "Nom"
        SYSTEM:
          type: string
          example: "Urine"
        TIME_ASPECT:
          type: string
          example: "Pt"
    ApiResponseError:
      type: "object"
      properties:
        status:
          type: "integer"
          format: "123"
          enum: [500]
        message:
          type: "string"
...
    ApiResponseCollectionNamesGet:
      type: array
      items:
        type: string
        example: "collection_1, collection_2"
    ApiResponseCollectionsDelete:
      type: object
      properties:
        message:
          type: string
          example: "The following tables were deleted: collection_1"
    ApiResponseCollectionDelete:
      type: object
      properties:
        message:
          type: string
          example: "Colection raw_medical_data_1's data was dropped.\nColection raw_medical_data_1 was deleted from the collection directory.\nMapping entries for raw_medical_data_1 were deleted."
...
    ApiResponseLoincPartValue:
      type: array
      items:
        type: string
        example: "PrThr"
    ApiResponseCollectionToLoincPartValue:
      type: array
      items:
        type: string
        example: "raw_medical_data"
...
    ApiResponseMetalevelLoincCodes:
      type: array
      items:
        type: object
        properties:
          loinc_code:
            type: string
            example: "5794-3"
          long_common_name:
            type: string
            example: "Hemoglobin [Presence] in Urine by Test strip"
    ApiResponseLoincCodeOperators:
      type: array
      items:
        type: object
        properties:
          code_opertor:
            type: string
            example: "ANY"
            enum: ["ANY", "ALL"]
    ApiResponseCollectionsToCompositeDescriptors:
      type: array
      items:
        type: object
        properties:
          collection_name:
            type: string
            example: "raw_medical_data"
...
    ApiResponseCollectionToLoinc:
      type: array
      items:
        type: string
        example: ["colelction1", "collection2"]
    ApiResponseCollectionToLoincNoCode:
      type: object
      properties:
        message:
          type: string
          default: "no loinc code" 
...